feat: add live preview to Projects admin editor and extract ProjectsSection
Extracts the grid/pagination logic from ProjectsPage into a reusable
ProjectsSection component, following the same page-vs-section split
already used across Home. This makes it embeddable in ProjectsEditor's
live preview, consistent with the other 6 CMS sections.
