+++ src/components/sections/ProjectsSection/index.ts
@@ -1,0 +1,1 @@
+export { ProjectsSection, type ProjectsSectionProps } from './ProjectsSection';
